% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{prepairr}
\alias{prepairr}
\alias{prepairr-package}
\title{\pkg{prepairr}: Interface to the automatic geometry repair tool 'prepair'}
\description{
Repair invalid geometry using constrained triangulation via the prepair
software library (Ledoux, Arroyo Ohori & Meijers, 2014,
\url{https://github.com/tudelft3d/prepair}).
}
\examples{
# create sf object with invalid geometry
x <- st_sfc(st_polygon(list(rbind(c(0, 0), c(0.5, 0), c(0.5, 0.5),
                                  c(0.5, 0), c(1, 0), c(1, 1), c(0, 1),
                                  c(0, 0)))))

# verify that geometry is invalid
suppressWarnings(st_is_valid(x))

# fix geometry using prepair
y <- st_prepair(x)

# verify that repaired geometry is valid
st_is_valid(y)

# plot the original and repaired geometries
par(mfrow = c(1, 2))
plot(st_geometry(x), main = "invalid geometry")
plot(st_geometry(y), main = "repaired geometry")

}
\references{
Ledoux H, Arroyo Ohori K, and Meijers M (2014) A triangulation-based
approach to automatically repair GIS polygons. Computers & Geosciences
66:121--131
}
\seealso{
st_prepair
}
